docs/build: don't unlink past files from dev-server

+++ webpack.config.ts
@@ -6,11 +6,14 @@ const cwd = process.cwd();
 const docs = path.join(cwd, 'docs');
 
 class NgAppDocsPlugin {
+	constructor(private isDev: boolean) { }
 	public apply(compiler: any) {
-		compiler.hooks.emit.tap(this.constructor.name, () => {
-			const files = fs.readdirSync(docs, { withFileTypes: true }).filter(x => x.isFile());
-			files.forEach(x => fs.unlinkSync(path.join(docs, x.name)));
-		});
+		if (this.isDev === false) {
+			compiler.hooks.emit.tap(this.constructor.name, () => {
+				const files = fs.readdirSync(docs, { withFileTypes: true }).filter(x => x.isFile());
+				files.forEach(x => fs.unlinkSync(path.join(docs, x.name)));
+			});
+		}
 		compiler.hooks.afterEmit.tap(this.constructor.name, () => {
 			fs.writeFileSync(path.join(docs, 'CNAME'), 'ng-app.js.org');
 		});
@@ -38,7 +41,7 @@ module.exports = (env = 'development') => {
 			template: '!!pug-loader?!docs/src/index.pug',
 			title: '@ledge/ng-app docs',
 		}),
-		new NgAppDocsPlugin(),
+		new NgAppDocsPlugin(env === 'development'),
 	);
 
 	return config;
